// src/shared/types/field-defenition.ts
type BaseField = {
  required?: boolean
  label: string
}

type InputField = BaseField & {
  fieldType: 'input'
  inputMode?: 'search' | 'text' | 'none' | 'email' | 'tel' | 'url' | 'numeric' | 'decimal'
  defaultValue?: string
}

type PasswordField = BaseField & {
  fieldType: 'password'
  defaultValue?: string
}

type TextareaField = BaseField & {
  fieldType: 'textarea'
  defaultValue?: string
}

export type FieldDefenition = InputField | PasswordField | TextareaField